Fix regression: not respecting interval in read_with_poll
The regression happened in the port to PL/pgSQL.

Division of integers on postgres yields integer results. Our unit conversion from
milliseconds to seconds didn't consider that, then returning wrong results.
This caused `pgmq.read_with_poll` to sleep 0s instead of `poll_interval_ms`
if `poll_interval_ms` was under 1000, using 100% of CPU.

The fix is simply converting `poll_interval_ms` to `numeric` before doing the
division, which produces a `numeric` result as originally intended.

+++ pgmq-extension/sql/pgmq--1.4.4--1.5.0.sql
@@ -97,7 +97,7 @@ BEGIN
       IF FOUND THEN
         RETURN;
       ELSE
-        PERFORM pg_sleep(poll_interval_ms / 1000);
+        PERFORM pg_sleep(poll_interval_ms::numeric / 1000);
       END IF;
     END LOOP;
 END;
